refactor: moved goal section component into its own file
