+++ frontend/src/components/ProgressCircle.tsx
@@ -1,0 +1,12 @@
+interface ProgressCircleProps {
+    percent: number
+}
+
+export function ProgressCircle(props: ProgressCircleProps) {
+    const {percent} = props
+    return (
+        <>
+        test
+        </>
+    )
+}
